Rename pioasm module due to name conflict
SConscript executes the script in the dumbest way possible: it loads the
script in and then calls eval on it.

This makes doing imports correctly impossible.

So when there's another pioasm module on the python module path, things
break, because python tries to import that module. Using a relative
import also doesn't work because of the previously mentioned dumbness:
python tries to import `SCons.Script.pioasm`.

The solution here is to change the name to something unique.